+++ projects/packages/forms/src/contact-form/class-contact-form-field.php
@@ -2076,9 +2076,10 @@ class Contact_Form_Field extends Contact_Form_Shortcode {
 		foreach ( $working_options as $option_index => $option ) {
 			$option_label  = Contact_Form_Plugin::strip_tags( $option['label'] );
 			$option_letter = Contact_Form_Plugin::strip_tags( $option['letter'] );
-			$image_block   = $option['image'];
+			$image_block   = $option['image'] ?? null;
+			$image_id      = ( is_array( $image_block ) && isset( $image_block['attrs'] ) && is_array( $image_block['attrs'] ) ) ? ( $image_block['attrs']['id'] ?? null ) : null;
 
-			$rendered_image_block = render_block( $image_block );
+			$rendered_image_block = is_array( $image_block ) ? render_block( $image_block ) : '';
 			// Remove any links from the rendered block
 			$rendered_image_block = preg_replace( '/<a[^>]*>(.*?)<\/a>/s', '$1', $rendered_image_block );
 
@@ -2104,7 +2105,7 @@ class Contact_Form_Field extends Contact_Form_Shortcode {
 					'label'      => $option_label,
 					'showLabels' => $show_labels,
 					'image'      => array(
-						'id'  => $image_block['attrs']['id'] ?? null,
+						'id'  => $image_id,
 						'src' => $image_src ?? null,
 					),
 				),