+++ projects/packages/publicize/_inc/components/form/upgrade-notice-customization.tsx
@@ -1,8 +1,9 @@
 import { getRedirectUrl } from '@automattic/jetpack-components';
 import { isSimpleSite } from '@automattic/jetpack-script-data';
-import { getSiteFragment } from '@automattic/jetpack-shared-extension-utils';
+import { useAnalytics, getSiteFragment } from '@automattic/jetpack-shared-extension-utils';
 import { Button, Flex, FlexItem, Notice } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
+import { useCallback } from 'react';
 
 /**
  * A notice for upgrading to a plan that supports per-network customization.
@@ -10,6 +11,12 @@ import { __ } from '@wordpress/i18n';
  * @return The UpgradeNoticeCustomization component.
  */
 export function UpgradeNoticeCustomization() {
+	const { recordEvent } = useAnalytics();
+
+	const onClickUpgrade = useCallback( () => {
+		recordEvent( 'jetpack_social_per_network_customization_upgrade_click' );
+	}, [ recordEvent ] );
+
 	if ( isSimpleSite() ) {
 		return null;
 	}
@@ -41,6 +48,7 @@ export function UpgradeNoticeCustomization() {
 						className="is-compact"
 						target="_blank"
 						rel="noopener noreferrer"
+						onClick={ onClickUpgrade }
 					>
 						{ __( 'Upgrade now', 'jetpack-publicize-pkg' ) }
 					</Button>