+++ projects/packages/publicize/_inc/hooks/use-per-network-customization/test/index.test.ts
@@ -31,6 +31,11 @@ jest.mock( '../../use-post-meta', () => ( {
 	} ) ),
 } ) );
 
+// Mock useAnalytics to avoid deep dependency chain
+jest.mock( '@automattic/jetpack-shared-extension-utils', () => ( {
+	useAnalytics: jest.fn( () => ( { recordEvent: jest.fn() } ) ),
+} ) );
+
 // Mock hasSocialPaidFeatures to return true by default
 jest.mock( '../../../utils', () => {
 	const actual = jest.requireActual( '../../../utils' );