+++ projects/plugins/jetpack/class.jetpack.php
@@ -1889,6 +1889,8 @@ class Jetpack {
 		 * @param bool true Should Twitter Card Meta tags be disabled. Default to true.
 		 */
 		if ( ! apply_filters( 'jetpack_disable_twitter_cards', false ) ) {
+			// @todo Remove this require once the deprecated Jetpack_Twitter_Cards wrapper has been removed.
+			// Twitter Cards functionality now lives in the jetpack-post-media package (Automattic\Jetpack\Post_Media\Twitter_Cards).
 			require_once JETPACK__PLUGIN_DIR . 'class.jetpack-twitter-cards.php';
 		}
 	}
@@ -5815,11 +5817,18 @@ endif;
 	}
 
 	/**
-	 * Returns a boolean for whether backups UI should be displayed or not.
+	 * Whether UI for backups should be displayed.
+	 *
+	 * On WPCom platforms this is gated on the backups-self-serve site feature.
+	 * On self-hosted Jetpack sites it falls back to the jetpack_show_backups filter.
 	 *
 	 * @return bool Should backups UI be displayed?
 	 */
 	public static function show_backups_ui() {
+		if ( ( new \Automattic\Jetpack\Status\Host() )->is_wpcom_platform() ) {
+			return function_exists( 'wpcom_site_has_feature' ) && wpcom_site_has_feature( 'backups-self-serve' );
+		}
+
 		/**
 		 * Whether UI for backups should be displayed.
 		 *
@@ -5828,6 +5837,22 @@ endif;
 		 * @param bool $show_backups Should UI for backups be displayed? True by default.
 		 */
 		return self::is_plugin_active( 'vaultpress/vaultpress.php' ) || apply_filters( 'jetpack_show_backups', true );
+	}
+
+	/**
+	 * Whether UI for security scanning should be displayed.
+	 *
+	 * On WPCom platforms this is gated on the scan-self-serve site feature.
+	 * On self-hosted Jetpack sites it always returns true.
+	 *
+	 * @return bool Should scan UI be displayed?
+	 */
+	public static function show_scan_ui() {
+		if ( ( new \Automattic\Jetpack\Status\Host() )->is_wpcom_platform() ) {
+			return function_exists( 'wpcom_site_has_feature' ) && wpcom_site_has_feature( 'scan-self-serve' );
+		}
+
+		return true;
 	}
 
 	/**