+++ projects/js-packages/social-previews/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.9] - 2026-02-19
+### Changed
+- Internal updates.
+
+## [3.0.8] - 2026-02-18
+### Added
+- Add notice for Threads link preview when no image is provided. [#47142]
+
+### Changed
+- Set `.repository.url` in `package.json` to the mirror repo rather than the monorepo. [#47149]
+
 ## [3.0.7] - 2026-02-16
 ### Changed
 - Update package dependencies. [#47099]
@@ -121,6 +132,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release after extracting from Calypso.
 
+[3.0.9]: https://github.com/Automattic/social-previews/compare/v3.0.8...v3.0.9
+[3.0.8]: https://github.com/Automattic/social-previews/compare/v3.0.7...v3.0.8
 [3.0.7]: https://github.com/Automattic/social-previews/compare/v3.0.6...v3.0.7
 [3.0.6]: https://github.com/Automattic/social-previews/compare/v3.0.5...v3.0.6
 [3.0.5]: https://github.com/Automattic/social-previews/compare/v3.0.4...v3.0.5